requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
coverage==4.5.4
Django==2.2.3
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.2
